Added config options for addtl extensions, pathing
Added optional features primarily aimed at supporting Laravel projects. Added the ability to generate tags in file types other than .html and the option to generate those tags from a directory other than the target file's. (For instance, /public.)

Only thing of concern to the existing package is that because "blade.php" is an extension with a '.' in it, I had to change `currentFilePathExtension` to `currentFileName` and used `endsWith` to make the comparisons. (Which is probably less efficient, but likely has a negligible impact overall.)